+++ fetch-ponyfill.typings.d.ts
@@ -1,0 +1,8 @@
+// Type definitions for fetch-ponyfill 4.1.0
+// Project: https://github.com/qubyte/fetch-ponyfill
+
+declare module 'fetch-ponyfill' {
+  function fetchPonyfill(options?: any): any;
+  namespace fetchPonyfill {}
+  export = fetchPonyfill;
+}